+++ mwes/mwe_11.jl
@@ -1,16 +1,14 @@
 using Pkg
-if ! ("Plots" ∈ keys(Pkg.project().dependencies))
+if ! ("ControlPlots" ∈ keys(Pkg.project().dependencies))
     using TestEnv; TestEnv.activate()
 end
-using Plots
-Plots.__init__()
+using ControlPlots
 using KiteModels, KitePodModels, KiteUtils
 
 set = deepcopy(se())
 kcu::KCU = KCU(set)
 kps3::KPS3 = KPS3(kcu)
 
-include("../examples/plot2d.jl")
 
 reltime = 0.0
 integrator = KiteModels.init_sim!(kps3, stiffness_factor=0.04)