+++ test/create_sys_image2.jl
@@ -3,19 +3,19 @@ if ! ("PackageCompiler" ∈ keys(Pkg.project().dependencies))
     @info "Installing PackageCompiler ..."
     Pkg.add("PackageCompiler")
 end
-if ! ("Plots" ∈ keys(Pkg.project().dependencies))
-    @info "Installing Plots ..."
-    Pkg.add("Plots")
+if ! ("ControlPlots" ∈ keys(Pkg.project().dependencies))
+    @info "Installing ControlPlots ..."
+    Pkg.add("ControlPlots")
 end
 @info "Loading packages ..."
-using KiteUtils, KitePodModels, KiteModels, Plots, AtmosphericModels, Reexport
+using KiteUtils, KitePodModels, KiteModels, ControlPlots
 using PackageCompiler
 
 @info "Creating sysimage ..."
 push!(LOAD_PATH,joinpath(pwd(),"src"))
 
 PackageCompiler.create_sysimage(
-    [:KiteUtils, :KitePodModels, :KiteModels, :Plots, :AtmosphericModels, :Reexport];
+    [:KiteUtils, :KitePodModels, :KiteModels, :ControlPlots];
     sysimage_path="kps-image_tmp.so",
     precompile_execution_file=joinpath("test", "test_for_precompile.jl")
 )